feat/#12-HomeUI 패키지 구성
