--- app/src/main/java/com/org/martall/Filterfragment.kt
@@ -1,4 +1,0 @@
-package com.org.martall
-
-class Filterfragment {
-}
