--- app/src/main/java/com/org/martall/OrderStatusFragment.kt
@@ -1,4 +1,0 @@
-package com.org.martall
-
-class OrderStatusFragment {
-}
