+++ app/src/main/java/com/org/martall/view/home/menu/MenuFragment.kt
@@ -1,0 +1,4 @@
+package com.org.martall.view.home.menu
+
+class MenuFragment {
+}
