+++ app/src/main/java/com/org/martall/view/home/notification/NotificationFragment.kt
@@ -1,0 +1,4 @@
+package com.org.martall.view.home.notification
+
+class NotificationFragment {
+}
