+++ app/src/main/java/com/org/martall/view/home/orderstatus/OrderStatusFragment.kt
@@ -1,0 +1,4 @@
+package com.org.martall.view.home.orderstatus
+
+class OrderStatusFragment {
+}
