+++ app/src/main/java/com/org/martall/view/home/profilesetting/ProfileSettingFragment.kt
@@ -1,0 +1,4 @@
+package com.org.martall.view.home.profilesetting
+
+class ProfileSettingFragment {
+}
