+++ app/src/main/java/com/org/martall/view/home/search/Filterfragment.kt
@@ -1,0 +1,4 @@
+package com.org.martall.view.home.search
+
+class Filterfragment {
+}
